Added a function to get BQ query result as a dataframe
This is kind of like the notebook magics provided by the bq client. This
requires setting up the parameters in a little more detail but with the
added advantage of a function which will return a dataframe for the
submitted query! This can then be used modularly in python scripts as
well as in a notebook.