Added a function to get the last element of a vector/list
Trying to `install.packages`, especially in docker containers where i do not have root privileges, raises error when R is unable to write in restricted directories. The last item in `.libPaths()` is set to `~/R/<R.version$platform>-library/<x.y>` where i can install packages without any superuser privileges. But getting the last item in R is not as easy as Python...and therefore the need for this function.