Added a function to generate polars dataframe from BQ
This is very useful for large dataframes! R's bigrquery downloads, and parses, the data but is unable to convert it to data.table in reasonable time (minutes) which hinders exploration of data. Therefore, decided to invest in using polars and duckdb for my explorations. Using duckdb and polars concurrently in a python session can be a big win.