Added a macro to generate a window around a particular date.
This is useful for comparing days around a particular date (holidays).  For instance, if we wish to compare 30 days around cybermonday this year compared to 30 days around cybermonday from previous year then we can use this macro to generate date axis that can be used for comparison.

This is especially useful when the difference between year over year anniversary dates is 53 weeks (371 days) instead of the customary 52 weeks (364 days).