Added a function to split a list into sublists
np.reshape requires the shape argument to be compatible with the original shape. This is because numpy does not allow [rj]agged arrays!  However, there is often a need to chunk the list into smaller sublists with the last sublist having fewer elements. Some array languages (for e.g., K and NIAL) provide an ability to create [rj]agged arrays!

Also, included an example of how to use the excellent Hypothesis library to ensure that this is correct! Hypothesis helped refine the assert in this function to use gt instead of ge. If stride is 0 then there would've been a division by zero error in the indexes calculation!