Modified to be able to compute proportions for compatible dicts.
Python's collections.Counter generates frequency distribution which can be used directly as an argument to the prop function. Learned this from how K handles fold/reduce for dicts! A table *is* a dictionary: the keys are the colnames and the length of values of items, which has to be same length for each of the cols, represent the rows of the table.